feat: frontend admin&user

+++ backend/src/routes/UserRouter.ts
@@ -171,7 +171,7 @@ userRouter.post(
  */
 userRouter.post(
   '/edit-admin',
-  asyncWrapper(async (req: any, res, next) => {
+  asyncWrapper(async (req, res, next) => {
     try {
       if (
         !checkIntegrity(req.body, ['id', 'userAccount', 'userName', 'email'])
@@ -198,7 +198,18 @@ userRouter.post(
   '/delete',
   asyncWrapper(async (req: any, res, next) => {
     try {
-      res.status(200).json(await Service.Delete(req.auth.id))
+      if (req.auth.group === Group.SUPER_ADMIN) {
+        res
+          .status(200)
+          .json(
+            new Restful(
+              CodeDictionary.DELETE_ERROR__USER_ADMIN,
+              '不能删除超级管理员账号',
+            ),
+          )
+      } else {
+        res.status(200).json(await Service.Delete(req.auth.id))
+      }
     } catch (e) {
       // TODO: 进行邮件提醒
       res.status(500).end()
@@ -233,6 +244,23 @@ userRouter.post(
       } else {
         res.status(200).json(await Service.Delete(id))
       }
+    } catch (e) {
+      // TODO: 进行邮件提醒
+      res.status(500).end()
+    }
+    next()
+  }),
+)
+
+/**
+ * 检查登陆状态
+ * @path /check
+ */
+userRouter.post(
+  '/check',
+  asyncWrapper(async (req, res, next) => {
+    try {
+      res.status(200).json(new Restful(CodeDictionary.SUCCESS, '已登录'))
     } catch (e) {
       // TODO: 进行邮件提醒
       res.status(500).end()
